+++ packages/pytest-simcore/src/pytest_simcore/aws_ec2_service.py
@@ -3,14 +3,17 @@
 # pylint: disable=unused-import
 
 import contextlib
+import datetime
 import random
-from collections.abc import AsyncIterator
+from collections.abc import AsyncIterator, Callable
 from typing import cast
 
 import aioboto3
 import pytest
 from aiobotocore.session import ClientCreatorContext
+from aws_library.ec2.models import EC2InstanceData, Resources
 from faker import Faker
+from pydantic import ByteSize
 from settings_library.ec2 import EC2Settings
 from types_aiobotocore_ec2.client import EC2Client
 
@@ -122,3 +125,25 @@ async def aws_ami_id(
     image = random.choice(images["Images"])  # noqa: S311
     assert "ImageId" in image
     return image["ImageId"]
+
+
+@pytest.fixture
+def fake_ec2_instance_data(faker: Faker) -> Callable[..., EC2InstanceData]:
+    def _creator(**overrides) -> EC2InstanceData:
+        return EC2InstanceData(
+            **(
+                {
+                    "launch_time": faker.date_time(tzinfo=datetime.timezone.utc),
+                    "id": faker.uuid4(),
+                    "aws_private_dns": f"ip-{faker.ipv4().replace('.', '-')}.ec2.internal",
+                    "aws_public_ip": faker.ipv4(),
+                    "type": faker.pystr(),
+                    "state": faker.pystr(),
+                    "resources": Resources(cpus=4.0, ram=ByteSize(1024 * 1024)),
+                    "tags": faker.pydict(allowed_types=(str,)),
+                }
+                | overrides
+            )
+        )
+
+    return _creator
